changet method getWeight()

+++ src/taskPackage/weightComparator.java
@@ -6,7 +6,7 @@ public class weightComparator implements Comparator<Pet>{
 
 	@Override
 	public int compare(Pet o1, Pet o2) {
-		return ((Integer) o1.getWeight()).compareTo((Integer) o2.getWeight());
+		return o1.getWeight().compareTo(o2.getWeight());
 	}
 
 }
